Create BE folder/ add ingest api
